% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-rolling-var-SS.R
\name{all_rolling_var_SS}
\alias{all_rolling_var_SS}
\title{All Steady State Search by Rolling Variance}
\usage{
all_rolling_var_SS(output, tolerance)
}
\arguments{
\item{output}{Matrix: A matrix of simulation results where rows represent species and columns represent time steps. Typically generated with \link{run_simulation}.}

\item{tolerance}{Numeric: The tolerance value used to determine the steady state based on the rolling variance.}
}
\value{
\code{Stable_gen}: A numeric value representing the generation (time step) where the system first reaches the steady state, or \code{NA} if no steady state is found.
}
\description{
All Steady State Search by Rolling Variance
}
\details{
This function searches for steady states in a simulation by analyzing the changes in the simulation results over time using rolling variance. The steady state is detected when the variance between successive time steps falls below a given tolerance threshold.
}
\examples{
# Example usage:

wd <- "~/Documents/LAB_ECO"

# Initial parameters
N_species = 2
C0 = 0.45
CN = 0.2
Diag_val = -0.5

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv" )
params <-  forge_data(N_species, seeds_path, C0, CN, Diag_val)

# Generate simulation
output <- run_simulation(N_species, params = params, times = 20)

tolerance = 0.0
all_rolling_var_SS(output, tolerance)
}
