% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-SS-find-and-save.R
\name{all_SS_find_and_save}
\alias{all_SS_find_and_save}
\title{SS_find_and_save_all Function}
\usage{
all_SS_find_and_save(uniqueID, output, tolerance, wd)
}
\arguments{
\item{uniqueID}{Character: A unique identifier for the simulation run.}

\item{output}{Matrix: Simulation results where rows represent species and columns represent time steps.}

\item{tolerance}{Numeric: Tolerance value for determining the steady state.}

\item{wd}{Character: Working directory where the results will be saved.

 @import utils}
}
\description{
This function searches for steady states in a simulation by analyzing the differences in the simulation results over time and saves the
generation where the steady state is reached.
}
\examples{
# Example usage:
wd = "~/Documents/LAB_ECO/Simulations"

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv" )
params <-  forge_data(N_species = 2, seeds_path, C0 = 0.45, CN =  0.2, Diag_val = -0.5)

# Simulate species interactions
output <- run_simulation(N_species = 2, params = params, times = 100)

# Generate ID
uniqueID <- forge_id(wd)

# All  Steady states
tolerance = 0.05
all_SS_find_and_save(uniqueID, output, tolerance, wd)

}
