% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge-directories.R
\name{forge_directories}
\alias{forge_directories}
\title{forge_directories function}
\usage{
forge_directories(wd)
}
\arguments{
\item{wd}{Character. The path to the working directory where the directories will be created.}
}
\value{
Prints messages to the console indicating which directories were created or if they already exist.
}
\description{
This function creates the necessary directories for simulations in the specified working directory.
}
\details{
The function constructs directories used for simulations, including "Outputs", "Parameters", and "Scan".
}
\examples{
# Example usage
wd <- "~/Documents/LAB_ECO"
forge_directories(wd)
}
