% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge-ID.R
\name{forge_ID}
\alias{forge_ID}
\title{Generate a Unique ID for Output Files}
\usage{
forge_ID(wd)
}
\arguments{
\item{wd}{Character: The working directory path where output files are stored. The function will check the "Outputs" subdirectory to generate a unique ID based on the existing files.}
}
\value{
A character string containing a unique ID to be used for saving files.
}
\description{
This function generates a unique ID by checking the output directory to ensure no duplicate IDs are created.
}
\details{
The function creates a unique ID by ensuring that the newly generated ID does not match any IDs in the output files within the specified working directory.
}
\examples{
# Example usage:
wd <- "~/Documents/LAB_ECO"
uniqueID <- forge_ID(wd)
print(uniqueID)
}
