% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge-seeds.R
\name{forge_seeds}
\alias{forge_seeds}
\title{forge_seeds Function}
\usage{
forge_seeds(n, min, max, wd)
}
\arguments{
\item{n}{Number of possible seeds generated.}

\item{min}{Minimum possible seed value.}

\item{max}{Maximum possible seed value.}

\item{wd}{Working directory where the seeds file will be saved.}
}
\description{
This function generates `n` number of random seeds, subsamples them, and selects one to generate each item for the `init_data` function.
}
\examples{
# Example usage
wd = "~/Documents/LAB_ECO/"
forge_seeds(n = 200, min = 2, max = 1000, wd)

}
