% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge-tolerance.R
\name{forge_tolerance}
\alias{forge_tolerance}
\title{Search for the best tolerance}
\usage{
forge_tolerance(paramSettings, seeds_path, wd, individual)
}
\arguments{
\item{paramSettings}{A list containing the following elements:
\itemize{
  \item \code{C0}: Numeric. The probability of no interaction between species (i.e., 0 interaction).
  \item \code{CN}: Numeric. The probability of negative interaction (<0) between species.
  \item \code{Diag_val}: Numeric. The diagonal values used in the interaction matrix.
  \item \code{tolerance}: Numeric. Tolerance value for determining the steady state. The tolerance is transformed in \code{individual_diff_SS} using \code{tolerance^2} to define the threshold for stability.
  \item \code{times}: Numeric. The number of generations to simulate.
}}

\item{seeds_path}{Character. The path to the seeds file used for the simulation.}

\item{wd}{Character. Working directory where the results will be saved.}

\item{individual}{Logical. If TRUE, the steady state search will be conducted per species rather than globally.}
}
\value{
A character string containing a unique ID to be used for saving files.
}
\description{
This function generates the necessary parameters for a simulation and compares different tolerance values to identify the best one for finding steady states.
}
\examples{
# Example usage:
wd <- "~/Documents/LAB_ECO/testing"
seeds_path <- file.path("~/Documents/LAB_ECO", "Seeds.tsv")

paramSettings <- list(N_species = 2, 
                      C0 = 0.45, 
                      CN = 0.2, 
                      Diag_val = -0.5, # Parameters for data generation
                      tolerance = 0.05, # Tolerance used for Steady State search
                      times = 100) # Number of generations   
individual = TRUE

# Function
forge_tolerance(paramSettings, seeds_path, wd, individual)
}
