% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual-rolling-window-SS.R
\name{individual_rolling_window_SS}
\alias{individual_rolling_window_SS}
\title{Rwindow_individual Function}
\usage{
individual_rolling_window_SS(uniqueID, output, tolerance, wd)
}
\arguments{
\item{uniqueID}{Character: A unique identifier for the simulation run.}

\item{output}{Matrix: Simulation results where rows represent species and columns represent time steps.}

\item{tolerance}{Numeric: Tolerance value for determining the steady state.}

\item{wd}{Character: Working directory where the results will be saved.}
}
\value{
Numeric vector \code{Stable_vec}, where each index corresponds to a species and contains the generation at which that species reaches the steady state.
}
\description{
This function analyzes simulation results to identify steady states by calculating rolling means and checking when the differences fall below a specified tolerance.
}
\details{
The function calculates the rolling mean of the output with a window size of \code{#generations * 0.1}.
It then computes the differences between consecutive rolling means and determines the generation where the difference is less than the specified tolerance.
}
\examples{
# Example usage:

wd = "~/Documents/LAB_ECO"
seeds_path <- file.path(wd, "Seeds.tsv")
params <- init_data(N_species = 2, seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

# Run simulation
times <- 100  # Define the number of generations
output <- run_simulation(N_species = 2, params = params, times = times)

# Generate unique ID
uniqueID <- forge_id(wd)

# Search for Individual steady state.
tolerance <- 0.00005
individual_rolling_window_SS(uniqueID, output, tolerance, wd)

}
