% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds-SS-reader.R
\name{load_individual_ss_rds}
\alias{load_individual_ss_rds}
\title{Load individual Steady State RDS file}
\usage{
load_individual_ss_rds(wd, uniqueID)
}
\arguments{
\item{wd}{Character. The working directory where the \code{Scan/SS_individual.rds} file is located.}

\item{uniqueID}{A unique identifier for the simulation.}
}
\value{
A data frame where each row represents a simulation and each column contains different information.
}
\description{
This function reads the individual RDS file and returns it as a table.
}
\examples{
# Example usage:
wd <- "~/Documents/LAB_ECO/testing"
uniqueID <- "1f52dc"
table <- load_individual_ss_rds(wd, uniqueID)

}
