% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-saver.R
\name{output_saver}
\alias{output_saver}
\title{Save Simulation Output to a File}
\usage{
output_saver(output, uniqueID, wd)
}
\arguments{
\item{output}{A data frame or matrix containing the simulation output to be saved. The columns represent generations, and rows represent species.}

\item{uniqueID}{Character. A unique identifier for the file, generated using the \code{generate_uniqueID} function.}

\item{wd}{Character. The path to the working directory where the file will be saved. The output file will be saved in the path: \code{wd/Outputs/O_uniqueID.tsv}.}
}
\value{
A message is printed to the console indicating the file's ID and save location.
}
\description{
This function saves the provided simulation output to a TSV file with a specified unique ID in the given working directory.
}
\details{
The function constructs the file path using the working directory (`wd`), the "Outputs" subdirectory, and the unique identifier (`uniqueID`). The output is saved in TSV format, where columns represent generations and rows represent species.
}
\examples{
wd <- "~/Documents/LAB_ECO"

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv")
params <- init_data(N_species = 2, seeds_path = seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

# Simulate species interactions
times <- 100  # Define the number of generations
output <- run_simulation(N_species = 2, params = params, times = times, norm = FALSE)

# Generate unique ID
uniqueID <- forge_id(wd)

# Save output to file
output_saver(output, uniqueID, wd)
}
