% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-line-reader.R
\name{params_line_reader}
\alias{params_line_reader}
\title{Read Parameters by Line}
\usage{
params_line_reader(uniqueID, wd)
}
\arguments{
\item{uniqueID}{Character. A unique ID string used for saving files. This ID should be generated previously using the \code{generate_uniqueID} function.}

\item{wd}{Character. The path to the working directory where files will be saved. The file will be saved at: \code{wd/Parameters/Seeds_save.tsv}.}
}
\value{
A list containing the following elements:
\itemize{
  \item \code{Interactions}: A matrix representing the interaction values between species.
  \item \code{Growths}: A numeric vector representing the growth rates of the simulated species.
  \item \code{Population}: A numeric vector representing the initial abundances of the simulated species.
  \item \code{Seeds}: A numeric vector representing the seeds used to generate the populations, interaction matrix, and growth rates.
}
}
\description{
This function reads the original parameters used for simulation. 
These parameters were created with the \link{forge_data} function and saved with the \link{params_line_saver} function.
}
\examples{
wd <- "~/Documents/LAB_ECO/Simulations"

# Generate parameters
# forge_seeds(n = 200, min = 2, max = 1000, wd)

seeds_path <- file.path(wd, "Seeds.tsv")
params <- init_data(N_species = 10, seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

# Generate unique ID
uniqueID <- forge_id(wd)

# Save parameters by line
params_line_saver(params, uniqueID, wd)

# Read original parameters with the function 
res <- params_line_reader(uniqueID, wd)

}
