% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-line-saver.R
\name{params_line_saver}
\alias{params_line_saver}
\title{Save Parameters by Lines}
\usage{
params_line_saver(params, uniqueID, wd)
}
\arguments{
\item{params}{List. A list generated with the \link{forge_data} function, containing the following elements:
\itemize{
  \item \code{Interactions}: A matrix representing the interaction values between species.
  \item \code{Growths}: A numeric vector representing the growth rates of the simulated species.
  \item \code{Population}: A numeric vector representing the initial abundances of the simulated species.
  \item \code{Seeds}: A numeric vector representing the seeds used to generate the populations, interaction matrix, and growth rates.
}}

\item{uniqueID}{Character. A unique ID string used for saving files. This ID should be generated previously using the \code{forge_ID} function.}

\item{wd}{Character. The path to the working directory where the file will be saved. The file will be saved at: `wd/Parameters/P_uniqueID.tsv`.}
}
\value{
A message indicating the path where the parameters file was stored.
}
\description{
This function saves simulation parameters line by line to a specified file for future reference.
}
\examples{
# Example usage:
wd <- "~/Documents/LAB_ECO/Simulations"

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv")
params <- init_data(N_species = 10, seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

# Generate unique ID
uniqueID <- forge_id(wd)

# Save parameters by line
params_line_saver(params, uniqueID, wd)

}
