% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-seed-reader.R
\name{params_seed_reader}
\alias{params_seed_reader}
\title{Regenerate Original Simulation Parameters from Seeds and Inputs}
\usage{
params_seed_reader(uniqueID, wd)
}
\arguments{
\item{uniqueID}{Character. A unique ID string identifying the specific simulation. This ID should be generated previously using the \code{generate_uniqueID} function.}

\item{wd}{Character. The working directory where the `Seed_save.tsv` file, generated by \link{params_seed_saver}, is located.}
}
\value{
A list containing the following elements:
\itemize{
  \item \code{Interactions}: A matrix representing interaction coefficients between species. It is regenerated by setting a previously used seed and following this procedure: 
    A binomial distribution with a success probability of `CN` is used to determine whether the interaction will be negative. 
    The resulting number is then multiplied by another binomial distribution with a success probability of `C0` to determine if the interaction will be null (0). 
    Finally, this product is multiplied by a uniform distribution between 0 and 1 to obtain the final interaction value. 
    The diagonal of the interaction matrix is filled with the value of `Diag_val` (user input).
    
  \item \code{Growths}: A numeric vector of the growth rates for the simulated species. It is regenerated by setting a seed and using 
    a uniform distribution between 0.001 and 1, representing the growth rate of each species.
        
  \item \code{Population}: A numeric vector of the initial abundances of the species. It is regenerated by setting a seed and using 
    a uniform distribution between 0.01 and 1, representing the initial populations of each species.
}
}
\description{
This function regenerates the original simulation parameters based on the seeds and inputs used to generate them, retrieved from previously saved files.
}
\details{
This function regenerates the original simulation parameters using the seed and input data from the `Seed_save.tsv` file. It returns the parameters in the form of a list containing the interaction matrix, growth rates, and initial population abundances.
}
\examples{
wd <- "~/Documents/LAB_ECO/Simulations"

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv")
params <- forge_data(N_species = 10, seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

# Generate unique ID
uniqueID <- forge_id(wd)

# Save parameters by seeds
params_seed_saver(N_species = 10,  C0 = 0.45, CN = 0.2, Diag_val = -0.5, params, uniqueID, wd)

# Read parameters by seed
params_seed_reader(uniqueID, wd)

}
