% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-seed-saver.R
\name{params_seed_saver}
\alias{params_seed_saver}
\title{Save Parameters by Seed for Future Regeneration}
\usage{
params_seed_saver(N_species, C0, CN, Diag_val, params, uniqueID, wd)
}
\arguments{
\item{N_species}{Integer. The number of species involved in the simulation.}

\item{C0}{Numeric. The probability of no interaction between species (i.e., zero interaction).}

\item{CN}{Numeric. The probability of negative interaction (<0) between species.}

\item{Diag_val}{Numeric vector. The diagonal values used in the interaction matrix.}

\item{params}{List. Parameters used in the simulation, specifically \code{params$Semilla} will be saved, containing:
\itemize{
  \item \code{Semilla[1]}: Seed for initial populations.
  \item \code{Semilla[2]}: Seed for the interaction matrix.
  \item \code{Semilla[3]}: Seed for species growth rates.
}}

\item{uniqueID}{Character. A unique ID string used for saving files. This ID should be generated previously using the \code{forge_ID}function.}

\item{wd}{Character. The path to the working directory where files will be saved. The file will be saved at: `wd/Parameters/Seeds_save.tsv`.}
}
\description{
This function saves specified parameters in a tabular format. The parameters include the number of species, interaction probabilities, diagonal values of the interaction matrix, and seed values used in different aspects of the simulation.
}
\details{
This function stores the provided parameters in a tabular format for future reference. The table includes the species count, interaction probabilities,
matrix diagonal values, and seed values.
}
\examples{
wd <- "~/Documents/LAB_ECO/Simulations"

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv")
params <- init_data(N_species = 10, seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

# Generate unique ID
uniqueID <- forge_id(wd)

# Save parameters by seeds
params_seed_saver(N_species = 10,  C0 = 0.45, CN = 0.2, Diag_val = -0.5, params, uniqueID, wd)

}
