% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Simulate_output Function}
\usage{
run_simulation(N_species, params, times, norm)
}
\arguments{
\item{N_species}{Numeric: Number of species involved in the simulation.}

\item{params}{List: A list including the following elements:
\itemize{
  \item Interactions: If \code{NULL}, a random interaction matrix is generated using \code{runif(n = n_species, min = 0, max = 1)}.
  \item Growths: Numeric vector representing the growth rates of the simulated species.
  \item Population: Numeric vector representing the initial abundances of the simulated species.
  \item Seeds: Numeric vector representing the seeds used to generate the interaction matrix, growth rates, and populations.
}}

\item{times}{Numeric: The number of generations to simulate.}

\item{norm}{Logical: Whether the output time series should be normalized to proportions (\code{norm = TRUE}) or returned as raw counts (\code{norm = FALSE}, default).}
}
\value{
The function returns the simulation results over time, typically in the form of a matrix or data frame.
}
\description{
This function simulates species interactions over a specified number of generations using miaSim::simulateGLV package
}
\examples{
wd <- "~/Documents/LAB_ECO"

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv")
params <- init_data(N_species = 2, seeds_path = seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

# Simulate species interactions
times <- 100  # Define the number of generations
output <- run_simulation(N_species = 2, params = params, times = times, norm = FALSE)
}
