% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteadyS_individual.R
\name{Rwindow_individual}
\alias{Rwindow_individual}
\title{Rwindow_individual Function}
\usage{
Rwindow_individual(uniqueID, output, tolerance, wd)
}
\arguments{
\item{uniqueID}{Character: A unique identifier for the simulation run.}

\item{output}{Matrix: Simulation results where rows represent species and columns represent time steps.}

\item{tolerance}{Numeric: Tolerance value for determining the steady state.}

\item{wd}{Character: Working directory where the results will be saved.}
}
\value{
Numeric vector \code{Stable_vec}, where each index corresponds to a species and contains the generation at which that species reaches the steady state.
}
\description{
This function analyzes simulation results to identify steady states by calculating rolling means and checking when the differences fall below a specified tolerance.
}
\details{
The function calculates the rolling mean of the output with a window size of \code{#generations * 0.1}.
It then computes the differences between consecutive rolling means and determines the generation where the difference is less than the specified tolerance.
}
\examples{
# Example usage:

wd <- "~/Documents/LAB_ECO"

# Define simulation parameters
N_species <- 2
C0 <- 0.45
CN <- 0.2
Diag_val <- -0.5

# Generate unique ID
uniqueID <- generate_uniqueID(wd)

# Generate simulation parameters
seeds_path <- file.path(wd, "Seeds.tsv")
params <- generate(N_species, seeds_path, C0, CN, Diag_val)

# Run simulation
times <- 20 # Define the number of generations
output <- Simulate_output(N_species, params = params, times = times, norm = FALSE)

# Calculate rolling window steady state
tolerance <- 0.05
steady_state <- Rwindow_individual(uniqueID, output, tolerance, wd)

}
