% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS-diff-means-all.R
\name{SS_diff_means_all}
\alias{SS_diff_means_all}
\title{SS_diff_means_all Function}
\usage{
SS_diff_means_all(output, tolerance)
}
\arguments{
\item{output}{Matrix: Simulation results where rows represent species and columns represent time steps.}

\item{tolerance}{Numeric: Tolerance value for determining the steady state. It is transformed using \code{log(tolerance^2)}.}
}
\value{
A list containing:
  \itemize{
     \item \code{Diff_means}: Numeric vector. Column means of the squared differences between successive time steps, transformed using \code{log}.
     \item \code{Tolerance}: Numeric. Original tolerance value used.
     \item \code{Transformed_tolerance}: Numeric. Tolerance value transformed using \code{log(tolerance^2)}, displayed with 5 decimal places.
     \item \code{Stable_gen}: Numeric or NA. Generation where the mean of the differences < tolerance transformed.
 }
}
\description{
This function searches for steady states in a simulation by analyzing the squared differences in the simulation results over time.
}
\examples{
# Example usage:

wd <- "~/Documents/LAB_ECO"

# Initial parameters
N_species = 2
C0 = 0.45
CN = 0.2
Diag_val = -0.5

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv" )
params <-  init_data(N_species, seeds_path, C0, CN, Diag_val)

# Generate simulation
times <- 20 # Define the number of generations
output <- Simulate_output(N_species, params = params, times = times, norm = FALSE)

tolerance = 0.05
result <- SS_diff_means_all(output, tolerance)

}
