% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer-output.R
\name{visualize_output}
\alias{visualize_output}
\title{Plot the output of all species}
\usage{
visualize_output(output)
}
\arguments{
\item{output}{A data frame or matrix containing the simulation output to be saved. The columns represent generations, and rows represent species.}
}
\value{
The resulting plot
}
\description{
This function plot the change of population over the time (generations)
}
\examples{
# Example usage
wd <- "~/Documents/LAB_ECO/testing"
uniqueID <- "bcfg45"

out_path <- paste(wd, "/Outputs/O_", uniqueID , ".tsv", sep = "") # output path
output <- as.matrix( data.table::fread(out_path, sep = "\t") )
visualize_output(output)
}
