% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge-directories.R
\name{forge_directories}
\alias{forge_directories}
\title{Generate required directories for the package's functions}
\usage{
forge_directories(wd)
}
\arguments{
\item{wd}{Character. The path to the working directory where the directories will be created.}
}
\value{
Prints messages to the console indicating which directories were created or if they already existed.
}
\description{
This function creates the necessary directories for simulations in the specified working directory.
}
\details{
The function creates directories used for simulations, including: 

\itemize{
  \item  \strong{\code{Outputs}}: Directory where the simulation matrix of population changes over time is saved.
  \item  \strong{\code{Parameters}}: Directory for saving the parameters used to generate the data. 
         Parameters can be saved either by line using \link{params_line_saver}, or by seed using \link{params_seed_saver}.
  \item  \strong{\code{Scan}}: Directory where the results of the steady state search algorithms are saved, including 
         global stable generations (using \link{all_SS_find_and_save}) or by individual species using \link{individual_SS_find_and_save}).
  \item  \strong{\code{Differences}}: Directory where the log-transformed differences \eqn{\log(t+1) - \log(t)} of the output matrix are saved.
}
}
\examples{
# Example usage
wd <- "~/Documents/LAB_ECO/Simulations"
forge_directories(wd)

}
