% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-data.R
\name{init_data}
\alias{init_data}
\title{Generate Parameters for Simulation}
\usage{
init_data(N_species, seeds_path, C0, CN, Diag_val)
}
\arguments{
\item{N_species}{Numeric. The number of species involved in the simulation.}

\item{seeds_path}{Character. The file path to a TSV file containing all possible seeds to sample. The seeds should be prepared
using the \link{forge_seeds} function.}

\item{C0}{Numeric. The probability of no interaction between species (i.e., 0 interaction).}

\item{CN}{Numeric. The probability of negative interaction (<0) between species.}

\item{Diag_val}{Numeric. The diagonal values used in the interaction matrix.}
}
\value{
A list containing the following elements:
  \itemize{
    \item \code{Interactions}: A matrix representing the interaction values between species.
    \item \code{Growths}: A numeric vector representing the growth rates of the simulated species.
    \item \code{Population}: A numeric vector representing the initial abundances of the simulated species.
    \item \code{Seeds}: A numeric vector representing the seeds used to generate the populations, interaction matrix, and growth rates.
  }
}
\description{
This function generates the necessary parameters for a simulation, which can then be used in the \link{run_simulation} function.
}
\details{
\itemize{
  \item For \strong{generating the populations}, a uniform distribution between 0.1 and 1 is used. This represents the proportion of each species in the medium.
  \item For \strong{generating the interaction matrix}, a binomial distribution with a success probability of `CN` is used to determine whether the interaction will be negative.
  
  The resulting number is then multiplied by another binomial distribution with a success probability of `C0` to determine if the interaction will be null (0). Finally, this product is multiplied by a uniform distribution between 0 and 1 to obtain the final interaction value.
  
  The diagonal of the interaction matrix is filled with the value of `Diag_val` (user input).
  \item For \strong{generating the growth rates}, a uniform distribution between 0.001 and 1 is utilized. This represents the growth rate of each species.
}
}
\examples{
# Example usage:
wd <- "~/Documents/LAB_ECO"
seeds_path <- file.path(wd, "Seeds.tsv")
params <- init_data(N_species = 2, seeds_path = seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)
print(params)
}
