% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-line-saver.R
\name{params_line_saver}
\alias{params_line_saver}
\title{params_line_saver Function}
\usage{
params_line_saver(params, uniqueID, wd)
}
\arguments{
\item{params}{List. A list containing elements generated by the \code{generate} function. This list should include the following elements:
\itemize{
   \item Population: Numeric vector representing the initial abundances of the simulated species.
   \item Interactions: If \code{NULL}, a random interaction matrix is generated using \code{runif(n = n_species, min = 0, max = 1)}.
   \item Growths: Numeric vector representing the growth rates of the simulated species.
   \item Seeds: A numeric vector representing the seeds used to generate the populations, interaction matrix, and growth rates.
}}

\item{uniqueID}{Character. A unique ID string used for saving files. This ID should be generated previously using the \code{generate_uniqueID} function.}

\item{wd}{Character. Path to the working directory where files will be saved. The file will be saved in the path: \code{wd/Parameters/P_uniqueID.tsv}.}
}
\value{
A list containing the generated ID and the path where the simulation file is stored.
}
\description{
This function saves parameters by lines.
}
\examples{
  wd <- "~/Documents/LAB_ECO"

  # Generate parameters for simulation
  seeds_path <- file.path(wd, "Seeds.tsv" )
  params <- init_data(N_species = 2, seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

  # Generate unique ID
  uniqueID <- forge_id(wd)

  # Save by lines
  params_line_saver(params, uniqueID, wd)

}
