% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed-saver.R
\name{seed_saver}
\alias{seed_saver}
\title{Seed_saver Function}
\usage{
seed_saver(N_species, C0, CN, Diag_val, params, uniqueID, wd)
}
\arguments{
\item{C0}{Numeric. Probability of no interaction between species (i.e., zero interaction).}

\item{CN}{Numeric. Probability of negative interaction (<0) between species.}

\item{Diag_val}{Numeric vector. Diagonal values used in the interaction matrix.}

\item{params}{List. Parameters used in the simulation, specifically \code{params$Semilla} containing:
\itemize{
  \item \code{Semilla[1]}: Seed for initial populations.
  \item \code{Semilla[2]}: Seed for the interaction matrix.
  \item \code{Semilla[3]}: Seed for species growth rates.
}}

\item{uniqueID}{Character. A unique ID string used for saving files. This ID should be generated previously using the \code{generate_uniqueID} function.}

\item{wd}{Character. Path to the working directory where files will be saved. The file will be saved in the path: \code{wd/Parameters/Seeds_save.tsv}.}

\item{N_specs}{Integer. Number of species involved in the simulation.}
}
\description{
This function saves specified parameters to a table format. The parameters include the number of species, interaction probabilities, diagonal values of the interaction matrix, and seed values used in different aspects of the simulation.
}
\details{
The function stores the provided parameters in a tabular format for future reference. The table includes species count, interaction probabilities,
matrix diagonal values, and seed values.
}
\examples{
 wd <- "~/Documents/LAB_ECO"

 # Generate parameters for simulation
 seeds_path <- file.path(wd, "Seeds.tsv" )
 params <- init_data(N_species = 2, seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

 # Generate unique ID
 uniqueID <- forge_ID(wd)

 # Save parameters using Seed_saver function
 seed_saver(N_specs = 2,  C0 = 0.45, CN = 0.2, Diag_val, params, uniqueID, wd)

}
